+++ src/main/java/frc/robot/subsystems/DriveTrainSubsystem.java
@@ -123,11 +123,11 @@ public class DriveTrainSubsystem extends SubsystemBase {
     drive.tankDrive(nextleftPower, nextRightPower, false);
   }
 
-  public int getLeftEncoder() {
+  public double getLeftEncoder() {
     return leftDriveFalconMain.getSelectedSensorPosition();
   }
 
-  public int getRightEncoder() {
+  public double getRightEncoder() {
     return rightDriveFalconMain.getSelectedSensorPosition();
   }
